[FEAT(ProfitCalculator)] 최종 수익 구하기

+++ src/main/kotlin/domain/User.kt
@@ -1,11 +1,11 @@
 package domain
 
-class User(val name: String, val cards: Cards, val betAmount: Int) {
+class User(val name: String, val cards: Cards, val betAmount: Double) {
 
     lateinit var gameResult: GameResult
 
     companion object {
         fun create(nameAndBetAmount: Pair<String, Int>, cards: Cards): User =
-            User(nameAndBetAmount.first, cards, nameAndBetAmount.second)
+            User(nameAndBetAmount.first, cards, nameAndBetAmount.second.toDouble())
     }
 }
